requirements.txt
beautifulsoup4==4.11.1
lxml==4.9.2
pandas==1.5.3
pyodbc==4.0.35
python-decouple==3.7
requests==2.28.2
SQLAlchemy==2.0.0
urllib3==1.26.14
